Начета разработака АИ для фулл пати Гуардов

+++ fuzzy/l2open/util/geometry/Vector/Vector2P.java
@@ -2,8 +2,8 @@ package l2open.util.geometry.Vector;
 
 public class Vector2P {
 
-    private double x;
-    private double y;
+    private int x;
+    private int y;
     private final Point2D origin;
     private final Point2D point;
     private final Point2D point_O2P;
@@ -17,16 +17,16 @@ public class Vector2P {
     }
 
     private Point2D calc() {
-        double c = 2;
-        double x = point_O2P.x * c - origin.x;
-        double y = point_O2P.y * c - origin.y;
+        int c = 2;
+        int x = point_O2P.x * c - origin.x;
+        int y = point_O2P.y * c - origin.y;
         return new Point2D(x, y);
     }
 
 
     public Point2D calcO(Point2D p1, Point2D p2){
-        double xo = (p1.x + p2.x) / 2;
-        double yo = (p1.y + p2.y) / 2;
+        int xo = (p1.x + p2.x) / 2;
+        int yo = (p1.y + p2.y) / 2;
         return new Point2D(xo, yo);
     }
 
@@ -39,19 +39,19 @@ public class Vector2P {
         return this.point;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
